not_use/requirements.txt
--find-links https://pypi.python.org/simple
# https://download.pytorch.org/whl/lts/1.8/torch_lts.html
py-trees
numpy>=1.26.4
stable-baselines3[extra]>=1.1.0
tensorboard
# tensorflow
future>=1.0.0
jsonpickle>=2.0.0
seaborn>=0.11.2
simplejson>=3.17.5
matplotlib>=3.4.3
pandas>=1.3.2
Pillow>=8.2.0
python-dateutil>=2.8.2
setuptools>=57.0.0
torch
torchvision
torchaudio

jinja2>=3.1.2

sortedcontainers==2.4.0
PyYAML==6.0.1

#torch-ac>=1.4.0
minigrid>=2.2.0
tensorboardX>=1.6
numpy>=1.3
gymnasium>=0.26
pygame>=2.5.2
array2gif>=1.0.4
bezier>=2023.7.28
pybts>=1.3.5
